Update issues after creating or setting dataset
