+++ homeassistant/components/otbr/__init__.py
@@ -2,92 +2,20 @@
 from __future__ import annotations
 
 import asyncio
-from collections.abc import Callable, Coroutine
-import dataclasses
-from functools import wraps
-from typing import Any, Concatenate, ParamSpec, TypeVar
 
 import aiohttp
 import python_otbr_api
-from python_otbr_api import tlv_parser
-from python_otbr_api.pskc import compute_pskc
 
 from homeassistant.components.thread import async_add_dataset
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady, HomeAssistantError
-from homeassistant.helpers import issue_registry as ir
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.typing import ConfigType
 
 from . import websocket_api
 from .const import DOMAIN
-from .util import get_allowed_channel
-
-_R = TypeVar("_R")
-_P = ParamSpec("_P")
-
-INSECURE_NETWORK_KEYS = (
-    # Thread web UI default
-    bytes.fromhex("00112233445566778899AABBCCDDEEFF"),
-)
-
-INSECURE_PASSPHRASES = (
-    # Thread web UI default
-    "j01Nme",
-    # Thread documentation default
-    "J01NME",
-)
-
-
-def _handle_otbr_error(
-    func: Callable[Concatenate[OTBRData, _P], Coroutine[Any, Any, _R]]
-) -> Callable[Concatenate[OTBRData, _P], Coroutine[Any, Any, _R]]:
-    """Handle OTBR errors."""
-
-    @wraps(func)
-    async def _func(self: OTBRData, *args: _P.args, **kwargs: _P.kwargs) -> _R:
-        try:
-            return await func(self, *args, **kwargs)
-        except python_otbr_api.OTBRError as exc:
-            raise HomeAssistantError("Failed to call OTBR API") from exc
-
-    return _func
-
-
-@dataclasses.dataclass
-class OTBRData:
-    """Container for OTBR data."""
-
-    url: str
-    api: python_otbr_api.OTBR
-
-    @_handle_otbr_error
-    async def set_enabled(self, enabled: bool) -> None:
-        """Enable or disable the router."""
-        return await self.api.set_enabled(enabled)
-
-    @_handle_otbr_error
-    async def get_active_dataset_tlvs(self) -> bytes | None:
-        """Get current active operational dataset in TLVS format, or None."""
-        return await self.api.get_active_dataset_tlvs()
-
-    @_handle_otbr_error
-    async def create_active_dataset(
-        self, dataset: python_otbr_api.OperationalDataSet
-    ) -> None:
-        """Create an active operational dataset."""
-        return await self.api.create_active_dataset(dataset)
-
-    @_handle_otbr_error
-    async def set_active_dataset_tlvs(self, dataset: bytes) -> None:
-        """Set current active operational dataset in TLVS format."""
-        await self.api.set_active_dataset_tlvs(dataset)
-
-    @_handle_otbr_error
-    async def get_extended_address(self) -> bytes:
-        """Get extended address (EUI-64)."""
-        return await self.api.get_extended_address()
+from .util import OTBRData, update_issues
 
 
 async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
@@ -96,100 +24,11 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
     return True
 
 
-async def _warn_on_channel_collision(
-    hass: HomeAssistant, entry: ConfigEntry, dataset_tlvs: bytes, otbrdata: OTBRData
-) -> None:
-    """Warn user if OTBR and ZHA attempt to use different channels."""
-
-    def delete_issue() -> None:
-        ir.async_delete_issue(
-            hass,
-            DOMAIN,
-            f"otbr_zha_channel_collision_{entry.entry_id}",
-        )
-
-    if (allowed_channel := await get_allowed_channel(hass, otbrdata.url)) is None:
-        delete_issue()
-        return
-
-    dataset = tlv_parser.parse_tlv(dataset_tlvs.hex())
-
-    if (channel_s := dataset.get(tlv_parser.MeshcopTLVType.CHANNEL)) is None:
-        delete_issue()
-        return
-    try:
-        channel = int(channel_s, 16)
-    except ValueError:
-        delete_issue()
-        return
-
-    if channel == allowed_channel:
-        delete_issue()
-        return
-
-    ir.async_create_issue(
-        hass,
-        DOMAIN,
-        f"otbr_zha_channel_collision_{entry.entry_id}",
-        is_fixable=False,
-        is_persistent=False,
-        severity=ir.IssueSeverity.WARNING,
-        translation_key="otbr_zha_channel_collision",
-        translation_placeholders={
-            "otbr_channel": str(channel),
-            "zha_channel": str(allowed_channel),
-        },
-    )
-
-
-def _warn_on_default_network_settings(
-    hass: HomeAssistant, entry: ConfigEntry, dataset_tlvs: bytes
-) -> None:
-    """Warn user if insecure default network settings are used."""
-    dataset = tlv_parser.parse_tlv(dataset_tlvs.hex())
-    insecure = False
-
-    if (
-        network_key := dataset.get(tlv_parser.MeshcopTLVType.NETWORKKEY)
-    ) is not None and bytes.fromhex(network_key) in INSECURE_NETWORK_KEYS:
-        insecure = True
-    if (
-        not insecure
-        and tlv_parser.MeshcopTLVType.EXTPANID in dataset
-        and tlv_parser.MeshcopTLVType.NETWORKNAME in dataset
-        and tlv_parser.MeshcopTLVType.PSKC in dataset
-    ):
-        ext_pan_id = dataset[tlv_parser.MeshcopTLVType.EXTPANID]
-        network_name = dataset[tlv_parser.MeshcopTLVType.NETWORKNAME]
-        pskc = bytes.fromhex(dataset[tlv_parser.MeshcopTLVType.PSKC])
-        for passphrase in INSECURE_PASSPHRASES:
-            if pskc == compute_pskc(ext_pan_id, network_name, passphrase):
-                insecure = True
-                break
-
-    if insecure:
-        ir.async_create_issue(
-            hass,
-            DOMAIN,
-            f"insecure_thread_network_{entry.entry_id}",
-            is_fixable=False,
-            is_persistent=False,
-            severity=ir.IssueSeverity.WARNING,
-            translation_key="insecure_thread_network",
-        )
-    else:
-        ir.async_delete_issue(
-            hass,
-            DOMAIN,
-            f"insecure_thread_network_{entry.entry_id}",
-        )
-
-
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up an Open Thread Border Router config entry."""
     api = python_otbr_api.OTBR(entry.data["url"], async_get_clientsession(hass), 10)
 
-    otbrdata = OTBRData(entry.data["url"], api)
+    otbrdata = OTBRData(entry.data["url"], api, entry.entry_id)
     try:
         dataset_tlvs = await otbrdata.get_active_dataset_tlvs()
     except (
@@ -199,8 +38,7 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     ) as err:
         raise ConfigEntryNotReady("Unable to connect") from err
     if dataset_tlvs:
-        await _warn_on_channel_collision(hass, entry, dataset_tlvs, otbrdata)
-        _warn_on_default_network_settings(hass, entry, dataset_tlvs)
+        await update_issues(hass, otbrdata, dataset_tlvs)
         await async_add_dataset(hass, DOMAIN, dataset_tlvs.hex())
 
     entry.async_on_unload(entry.add_update_listener(async_reload_entry))
